Fix manila update, upgrade, and FFU tasks
Manila has a similar service architecture to Cinder and the
latter has received several fixes for update, upgrade, and FFU
tasks that are missing in Manila.

Align these tasks in the Manila services with Cinder versions
of the same.

Note: I resolved conflicts in
docker/services/pacemaker/manila-share.yaml
due to the big refactor patch [1] in master that was not [2]
backported to stable/queens.  We need this backport nonetheless since
upgrades and FFU do not work for manila on stable/queens.

Also, this backport also incorporates a fix to a typo in master [3]
discovered after the original for this cherry-pick had merged there.

[1] https://review.openstack.org/#/c/562294
[2] https://review.openstack.org/#/c/576520
[3] https://review.openstack.org/#/c/588407/

Closes-Bug: #1784032

Change-Id: Ifae95bae859a301e6bcf67910cd9140f51584bb3
(cherry picked from commit 0811ccbc543a8cac991b10d8cea5498b9616e40f)