Do not target Redis over TLS when using HA Redis profile
In Pike, Redis doesn't support being accessed over a TLS tunnel when
configured for HA deployments.

Fix the Redis service configuration by force-disabling the TLS
bits in the generated Redis config, as well as in the Redis clients
configuration (gnocchi_api and HAProxy). The hiera keys are
defined in service_config_settings to be compatible with composable HA.

Change-Id: I1602f2b90d7166df05e9a34057064010d6a31afe
Closes-Bug: 1784205
Depends-On: I3b50ebf50046d33d49fbb555c2ee3bc6b3c2eaf9