[Pike only] ManilaBackendGeneric nooped to avoid deprecated error.
ManilaBackendGeneric service has been deprecated
in Pike release. But if pike version prior to the
service deprecation was installed and then a minor
update is performed to latest Pike an error will
be returned as the service is not found anymore.
This patch sets the service to OS::Heat::None to
bypass this issue.

Closes-Bug: #1784624
Change-Id: Ia3a55133f9f11eb8678eeff7a4062d19318aa217