ovn: Fix minor update failure with OVN db pacemaker HA resource
If an OVN deployment doesn't have this fix [1] in the ovn-northd docker image,
the if the image is not tagged with pcmklatest (see [2]), then during
step 5 of minor update of controllers (on the bootstrap node) the
'pacemaker_resource' ansible module fails for the "state: disable" task.

It is observed in my testing that 'pacemaker_resource - state: disable' ansible
tasks fails if the resource is already in stopped state. It also fails
for the task 'pacemaker_resource - state: enable' if the resource is already
enabled. This probably needs a fix in 'pacemaker_resource' ansible module.

To fix the stack update failure issue with OVN, this patch removes these
tasks - disable and enable of ovn-dbs-bundle before and after running the
command "pcs resource bundle update ovn-dbs-bundle container
image={{ovn_dbs_docker_image_latest}}". Since updating the bundle triggers
the resource restart, probably we don't need to explicitly disable and
enable the resource.

[1] - https://github.com/openvswitch/ovs/commit/9b46d3f609abf8ad532efb0820753f3a920e2213#diff-9c26550f1dff9ce66cf9c3c510b47864
[2] - See commit with change id - Idbda3bedff57376b74269ab3470d2324b804ffd4

Change-Id: I3f47fd019ca6b2ccb378c97751a11cb12ffed521
Closes-bug: #1784863