Switch deployment_source_hosts default to "Undercloud"
"undercloud" host is too opinionated and hostnames can change. We should
rather apply the tasks to the Undercloud HostGroup, which contains one
host for now: the actual undercloud hostname.

So this patch switches "undercloud" to "Undercloud" so when the hostname
isn't "undercloud", the external tasks will run correctly on this host.

Change-Id: I7200f930387406e6cc8e6fee6d5278768074c892
Closes-Bug: #1784910