Move the shebang to /usr/bin/python
In LP#1785245 we made sure to move the shebang script back to
/bin/python. rpm packaging inspect script shebang lines and add Requires
to the rpm. repoclosure checks to make sure that a given repo
(collection of rpm files) has all necessary Requires to be able to
install.

Let's move the shebang to /usr/bin/python which exists and does
not cause repoclosure issues:
[root@compute-0 ~]# docker exec -it nova_compute sh -c 'rpm -qf /usr/bin/python'
python-2.7.5-69.el7_5.x86_64

Change-Id: Iac4236bebab609e78b089b809a433d62c60e28c2
Reported-By: Jon Schlueter <jschluet@redhat.com>
Related-Bug: #1785245