Add cleanup service for neutron ovs bridges
Neutron used to run a neutron-ovs-cleanup service on startup and
shutdown to ensure clean configuration. Container specific version of
the cleanup service is needed as we are no longer installing neutron
packages in the overcloud image and the ordering dependencies of the
legacy services are not suitable for containerized neutron deployments.

Closes-Bug: #1786311

Change-Id: Iddd191f06b0a7cadb463abec7ef152c89fddb3cd
(cherry picked from commit 81265737187f32eef2b654a901ed74818873eb88)