Remove External network from DVR related configuration
DVR doesn't imply the "External" networks, simply some type of
connection that is being used for floating IP traffic. This patch
removes the External network from the DVR role and "left overs" from
environment files. It also corrects the multiple-nics version of the
templates so that the ComputeDVR role is attached to the external
bridge with no IP.

Co-Authored-By: Dan Sneddon <dsneddon@redhat.com>

Change-Id: Ia599e01dbefe4e4c752b7d4c1c7e5682963101f7