Add OctaviaEventStreamerDriver and change default
octavia::health_manager::event_streamer_driver is forcely set to
'queue_event_streamer' with no option to change it. The event streamer
specifies which driver to use for syncing Octavia and Neutron LBaaS
databases. It is highly recommended to disable if one doesn't need to
sync the database or is running Octavia in standalone mode by setting
event_streamer_driver to noop_event_streamer.

Change-Id: I400ce9a059aba195cdd987ba51360ef3d7f258f3
Closes-Bug: #1787608