Bind mount the database client settings in sahara_db_sync
Apparently not doing this break the bootstrap container in some cases
(for example when TLS everywhere is enabled). This case is not really
supported by Sahara right now, but better fix it in advance.

More details about this change are available in the similar patches
that landed for other components:
- Cinder: https://review.openstack.org/539498
- Manila: https://review.openstack.org/594801

Change-Id: Iab8ad50f4397ee9809f50d1474026d5ff8a6972c