Recreate cinder LVM loopback device on startup
Add a systemd service that creates the loopback device required by
cinder's iSCSI backend on system startup.

This patch also consolidates the host_prep_tasks for the HA and non-HA
versions of the cinder-volume service. The list of tasks is identical,
and rather than repeating it in each template, the tasks are defined
once in cinder-common.yaml.

Closes-Bug: #1581092
Change-Id: Icc04003a9e90b66720d968c6c8f1c687156b677e