Run online data migrations
We always run DB sync in deploy_tasks, ensuring that the database is
up to date. We should follow up with online data migrations
too.

Doing this via docker_config has 2 purposes:

* We can easily ensure this happens in a container with the right
  config files mounted.

* We can even apply this via a minor update. This is important because
  we'll have to backport this all the way to Pike and apply it there
  using Pike containers, before upgrading to Queens containers.

There's an additional issue to consider: In Puppet service variant we
ran the online migrations for release X before upgrading to X+1, but
the proposed Docker variant migrations for X run with upgrade to
X. This means that when switching from non-containerized to
containerized, we'll need to run migrations twice, to correctly switch
between the aforementioned approaches.

Change-Id: I2eb6c7c42d7e7aea4a78a892790e42bc5371f792
Closes-Bug: #1790474