Add a gateway IP to the Management net in network_data.yaml
In several places in our documentation, we discuss enabling
the Management network as a default route in some cases.
However, we don't set this value by default. Setting a
gateway IP in the example versions of network_data.yaml
would help to demonstrate this, and would create a good
default for testing.

Change-Id: Ibbb2a403f0733a016f1f49b4acdbdcfe7b9ab5a3