Fix openshift new node detection
For openshift deployment it is important to know which nodes are new to
the cluster so that can be added to the appropriate groups. The new node
detection was broken due to the recent switch to containerized
undercloud and move to openshift 3.10. This commit fixes the issues by
relying on the systemd status rather than docker directly.

Change-Id: Ica13d1aaee259d831d32698466ef8971e526483a