Clean up previous osa inventory dir before deployment
Since we're now pointing to an inventory directory instead of inventory
file, it is possible that, in certain condition such as when modifying
the roles, the inventory files from the previous deployment conflict
with current one. To prevent any issue, let's delete the inventory
directory prior to the deployment so that only the newly generated
inventory files are present.

Change-Id: I3e8d0d369c2ed5cac167f2958eddcb29f09eafa5