Add simple validation that OpenShift is deployed
Currently, we have no test to check if we actually deploy OpenShift
at the end of the scenario009 job. This can easily lead to false
positives from this job if some deploy step fails silently.

This patch adds two very basic commands to check if the OpenShift
cluster is actually operational.

Depends-On: Ibe94a58e0083b75e2f4cc904be6c17b0633d0293
Depends-On: Id24a7a0d9d34df4fd52ee1f99da61b873e5b870a
Depends-On: Ie21cc0dfed28353a13d419ff38cea82e65166221

Change-Id: Iee477ea736aff4290b703cd70258966819228c4a