check mode: docker_puppet_tasks
Adds check mode support for docker_puppet_tasks.

Since it's not possible to reliably determine what these tasks do, we
can't actually run them to get an idea of what might be changed. We can
however show the diff of the json file to get an idea of what would be
run.

Change-Id: I19e8bc9eb93d8acc8ee7d737770f9cc7e63f7a27