[Queens] Don't stop docker on config changes
This will allow us to take advantage of the Docker live restore
option, and not restart all containers during minor update on small
docker config changes. The containers will be restarted only in case
Docker RPM is updated during the minor update.

If we ship a config change which wouldn't work well with live restore,
we'll need to re-enable the restarts on config change. Docker
documentation does not specify which config options are safe to change
and which aren't.

Change-Id: I97cd28a7e1dfc1bec943754fbf42519b6f563da6
Closes-Bug: #1777146