mysql: do not overwrite password file during docker-puppet
During a stack update, when docker-puppet regenerates configs files for the
mysql service, the root mysql passwords may change. Mysql has to update its
internal state (e.g. password in mysql DB) to reflect the change, but this
only happens when paunch restarts mysql; and the old password it required
to until the change is applied.

For such services, update the config hash to notify paunch that a restart is
needed, but do not update the password file in docker-puppet and let the
service's containers regenerate it instead.

Change-Id: I5bdbc89897a6dcd5bd57f2132e2acf99702b28ea
Partial-Bug: #1792416
(cherry picked from commit 8e67ec833173920ac60b5548a711885a4d28e16f)