check mode: paunch configs
Adds initial check mode support for the paunch container startup
configuration and kolla config files. This cleans up the formatting of
the generated files so that the diff shown duing check mode with --diff
is useful.

We can't actually run paunch during check mode as it doesn't yet have
any support for a dry run mode.

Change-Id: I9add7b9fda50847c111e91735bd55a1ddf32f696