Undercloud - Restart keepalived on update
instack-undercloud had a workaround (30-reload-keepalived)
in place to always restart keepalived on install/upgrade.
This is required to ensure VIP's are present in case the
network config was changed and os-net-config restarts
the network interface. When containerizing the undercloud
this workaround was missed.

This change adds a similar workaround. A pre_deploy
NodeExtraconfig script will restart the keepalived
container when the undercloud installer is (re-)run.

NOTE: We can remove this workaround once keepalived
      v2.0.6 or later is available.

Closes-Bug: #1791238
Change-Id: I8cada7be57cd50c54ca5f2f38ec010062512ae06