Check if openstack-glance-registry is enabled before stopping it.
If glance-registry is defined in the service
parameters, then during upgrade, it will try
to stop it. However, as the service is not
enabled the service stop task will fail.
This patch checks if the service is enabled
before stopping it. And covers the case in
which the service might be disabled but still
running.

Closes-Bug: #1793557
Change-Id: I21d17c05f555b25f709f314f6a19a011a3d10a8d