Add a dual-bond example NIC config template
This change adds a 2-bonds-with-vlans example template which
demonstrates the use of two Linux bonds. This template will
place the 'Tenant*' networks on a bond with an OVS bridge.
Other networks will be placed as VLANs on the Linux bond
without a bridge. There is special handling for the Tenant
network on DPDK-enabled Compute nodes.

Change-Id: I9277c0e6a1267392943214eb5fe55509f7956fbc