Make openshift-master service idempotent
The openshift-master service will fail any time it is used with a
stack update. This is because the openshift_upgrade var is not
defined, but is checked whenever tripleo_stack_action == 'UPDATE'.

This patch adds a check for openshift_upgrade being defined before
checking if it is True.

Closes-Bug: 1794824
Change-Id: I3a598724154a3242b777eefed9304300c45d8c29
(cherry picked from commit d295cd94b82218f62c4b564f0a3a76cc04cba549)