[Ocata/Pike] Pass DeployIdentifier in upgrade tasks.
The UpdateIdentifier timestamp does not get updated when
running the overcloud deploy command, which is the one used
for upgrading. As a consequence, if the upgrade_tasks managed
to run once they won't be executed again by heat when re-issueing
the upgrade command, as heat considers that the resource did
not change. For this reason, we need to pass DeployIdentifier
which will contain a different timestamp in each execution and
will allow the upgrade_tasks to be executed in the overcloud nodes.

Change-Id: I49cecd8c793519185096fd99713521961022aec9
Closes-Bug: #1795790
Resolves: rhbz#1634034