Tag container image prepare tasks to allow running them for updates/upgrades
Updates/upgrades workflow must not run during `upgrade prepare` or
`upgrade run`, but during `upgrade run` we need to have the images
available. So the intention is to run `external-upgrade run --tags
container_image_prepare` between `upgrade prepare` and `upgrade run`.

The situation is analogical for `update` and `external-update`
commands.

Change-Id: I49de9a41c62204ab7cd835fec6dab8d59b054948
Closes-Bug: #1795881