Transitioning to HAProxy 1.8
Haproxy 1.8 brings in a specific change that breaks us:
It removes the haproxy-systemd-wrapper which
we use in order to be able to reload the config file without
restarting the whole container (important in TLS scenarios).

We fix this by calling the haproxy binary directly and
using the master-worker mode (-Ws) which allows to receive
a SIGUSR2 command which will then reload the config for
all the workers. It should also not background.

This commit keeps backward compatibility with current HAProxy
to ease the transition to new HAProxy.

Co-Authored-By: Damien Ciabrini <dciabrin@redhat.com>

Change-Id: I93943efefa22b9107c85f9f5e0bd4c3c1ab867ed
(cherry picked from commit e26ef65e5093f4f342b888d51dee925605931909)