Use different base virtual_router_id on openshift
With the default setting, the keepalived that we deploy on the master
node collides with the one that is setup on the undercloud. We simply
need to use a different virtual_router_id_base to prevent
virtual_router_id collision.

Change-Id: I92ef081a111f93ddce4ec42400bcb648b7f7def0
(cherry picked from commit 429bc3ca702c4bb68062a1d9e653d528246d0950)