Verify nova api migration finished before start placement
There is a deployment race where nova-placement fails to start if
the nova api db migration have not finished before starting it.
We start nova placement early to make sure it is up before the
nova-compute services get started. Since in HA scenario there is
no sync in between the nodes on the current worked deployment step
we might have the situation that the placement service gets started
on C1/2 when the nova api db sync is not yet finished on C0.

We have two possibilities:
1) start placement later and verify that nova-computes recover correct
2) verify that db migration on nova_api db finished before start nova-
placement on the controllers

This is a way to address 2) and should be the less invasive change as
we make sure placement gets started after the nova api db sync finished
and does not move the start to a complete different step which might
introduce any unforeseen behavior.

Change-Id: Ifc4346658cea5355bf89f6d7023ed846d9162b3e