[Queens/Pike] Re-registry with new rhel_reg parameters on upgrade.
Currently, when the overcloud nodes were registered in Satellite
or portal during deployment via extraconfig/pre_deploy/rhel-registration
and an upgrade is performed passing different values for rhel_reg_repos
these changes are not taken into account. As a consequence, we end up
running the upgrade steps with the previous repositories configured
during deployment.

This patch allows us to re-register using the --force option using
the new values inserted at the pre-upgrade step. To allow this, we
make use of the RHELRegistrationAction parameter to change the
action in which the RHELRegistrationDeployment resource is run.
By setting it to UPDATE, the SoftwareDeployment resource will
be triggered during the stack update which takes place in the
upgrade. Passing the force option will force the re-execution
of rhel-registration script even though it was executed previously
during deployment (bypassing the /mnt/state/var/lib/rhsm/rhsm.ok
guard.

Change-Id: I51a1dad5f422756d0a8c4deff81c170be7d944d3
Related-Bug: #1797138