[Ocata/Pike/Queens] Make rhel-registration scripts location absolute.
This patch modifies the script files location to be executed
during the SoftwareDeployment resources in rhel-registration.yaml.
The main purpose of this is to ensure we are running the right
script content related to a branch, for example, after an upgrade.

Change-Id: If77acb78c63511054e17d553982e19e748c88992
Related-Bug: #1797138
(cherry picked from commit 91746c290be5fc374ad2389694779a7b57351c81)