Sharing hypervisor /var/lib/ironic with ironic-conductor container
While puppet-config generation, ironic-conductor will call ironic::pxe
to copy files and create directories required for tftpboot and httpboot
to /varlib/ironic on container.

But /var/lib/ironic of hypervisor is not being shared with ironic-conductor,
the changes that were done on container are not present on hypervisor.

For ironic-inspector the previously generated configs are required to move
forward and apply its own puppet-configs.

In order to make the configs generated by ironic-conductor available to
ironic-inspector, hypervisor /var/lib/ironic must be shared with
ironic-conductor container as well while generating puppet-config.

Change-Id: I5ee1a78ea7b7681e0dc1387fc5af598919ecb397
Closes-Bug: 1795776