Ensure octavia flavor gets configured if composable octavia
The octavia service templates configure the octavia flavor through the
controller-worker is deployed on a node other than the controller. Since
this uses docker_puppet_tasks, it is only run on the bootstrap node
which at the moment means the first controller. This patch moves the
flavor creation to the docker API template to make sure the flavor gets
created when deploying worker components in an non-standard role.

Closes-Bug: #1797443

Change-Id: Id823672b82ea60f355d1f01068f802567d687da5