Ensure the needed openshift resources are set
We expect the the Keepalived and HAproxy services to be deployed on the
OpenShift master nodes, let's require them in the openshift heat
environment file. This prevents an issue when the docker-ha environment
is loaded because it would redefine these resources.

Change-Id: I57a7ea854bd8db4e20af1a608a6937604c0e3bd2
(cherry picked from commit 633379f77900995a71bdd7e85d73f975cd74c222)