Invoke redhat-subscription role during upgrade.
Before performing an upgrade, the new repositories and
activation key should be configured. Rhsm is the service
in charge of performing the overcloud nodes registration,
so we need to enable the new repository/parameters in the
upgrade_tasks to point to the right repositories during
upgrade.
Also, after [0] got merged, the registration is being forced
via --force in case it fails for the first time. Plus,
there is also a new option that allows us to enable only
the repositories passed in rhsm_repos variable, disabling
what is not included and was enabled. We need to pass
this variable set to true during upgrades so we disable
the latest version repositories.

[0] - https://review.openstack.org/#/c/615689/

Change-Id: I10fb84baa67275ed1f01fbb5b1236a284f2c763e
Related-Bug: #1797138